addressed gabes changes
