+++ firebase/firestore/invite_admin.ts
@@ -28,3 +28,18 @@ export const deleteAdmin = async (adminId: string) => {
     throw e;
   }
 };
+
+/**
+ * Checks if email is in the invited emails table
+ */
+export const checkAdminActivationStatus = async (
+  email: string
+): Promise<boolean> => {
+  try {
+    const qs = await adminInvitesCollection.where("email", "==", email).get();
+    return qs.docs.length > 0;
+  } catch (e) {
+    console.error(e);
+    throw e;
+  }
+};
